fix: contentEditable text encoding
